Move _frontend_request to frontendcomm
