+++ spyder_kernels/comms/frontendcomm.py
@@ -17,8 +17,9 @@ import time
 from jupyter_client.localinterfaces import localhost
 from tornado import ioloop
 import zmq
+from IPython.core.getipython import get_ipython
 
-from spyder_kernels.comms.commbase import CommBase
+from spyder_kernels.comms.commbase import CommBase, CommError
 from spyder_kernels.py3compat import TimeoutError, PY2
 
 
@@ -180,3 +181,16 @@ class FrontendComm(CommBase):
                 comm._msg_callback(msg)
         comm.handle_msg = handle_msg
         super(FrontendComm, self)._register_comm(comm)
+
+
+def _frontend_request(blocking=True):
+    """
+    Send a request to the frontend.
+
+    If blocking is true, The return value will be returned.
+    """
+    if not get_ipython().kernel.frontend_comm.is_open():
+        raise CommError("Can't make a request to a closed comm")
+    # Get a reply from the last frontend to have sent a message
+    return get_ipython().kernel.frontend_call(
+        blocking=blocking, broadcast=False)
